Modified example functions to work with 64-bit (long long) numbers.

+++ assembly/scanning.c
@@ -2,8 +2,8 @@
 
 int main()
 {
-	int a;
-	fscanf(stdin,"%d", &a);
+	long long int a;
+	fscanf(stdin,"%lld", &a);
 
-	fprintf(stderr, "%d\n", a);
+	fprintf(stderr, "%lld\n", a);
 }
